refactor(headless-future): centralize endpoint keys and remove config double-source
