+++ packages/headless-future/src/core/internal/api/base-facade/endpoint-keys.ts
@@ -1,0 +1,7 @@
+export const endpointKeys = {
+  conversation: 'conversation',
+} as const;
+
+export type EndpointKey = (typeof endpointKeys)[keyof typeof endpointKeys];
+
+export const conversationEndpointKey = endpointKeys.conversation;
